app/backend/requirements.txt
# ProbeScout backend. Implementation design §0c, §10.
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pyyaml>=6.0
openai>=1.0.0
pydantic>=2.0.0
